Update
- NGUI 3.7.1
- Removed plugins.

+++ Space CUBEs Project/Assets/Little Byte/Scripts/Data/SaveData/SaveData.cs
@@ -262,7 +262,10 @@ namespace LittleByte.Data
         /// </summary>
         public static void DeleteAll()
         {
-            Directory.Delete(DataPath, true);
+            if (Directory.Exists(DataPath))
+            {
+                Directory.Delete(DataPath, true);
+            }
             Directory.CreateDirectory(DataPath);
         }
